PHX-12352 Refactored Events and Regions into separate folders (too many classes in a single file) Added MultiEventToMultiGroupWorkflow and SpecificationEndpoints Added MultiEventToMultiGroup.md workflow document placeholder
