+++ sampleCode/CSharp/ConsoleApp/Endpoints/Events/MarginType.cs
@@ -1,0 +1,7 @@
+﻿namespace ConsoleApp.Endpoints.Events;
+
+public enum MarginType
+{
+    ProducerPrice,
+    PurchaserPrice
+}
