+++ sampleCode/CSharp/ConsoleApp/Endpoints/Events/SpendingPatternValueType.cs
@@ -1,0 +1,7 @@
+﻿namespace ConsoleApp.Endpoints.Events;
+
+public enum SpendingPatternValueType
+{
+    IntermediateExpenditure,
+    Output
+}
